requirements.txt
scikit-learn
numpy
scipy
fastapi
uvicorn[standard]
mlflow-skinny

